requirements.txt
feedparser>=6.0
PyYAML>=6.0
reportlab>=4.0
